Playerの挙動をフリックとスライドで制御するようにした

+++ Assets/Scripts/Charactor/Player/SlideCheckToAttack.cs
@@ -6,7 +6,7 @@ public class SlideCheckToAttack
 {
     AttackClass m_attack = new AttackClass();
 
-    public void IsSlide(GameObject get, int dir)
+    public void IsSlide(GameObject get, float dir)
     {
         PlayerController player = get.GetComponent<PlayerController>();
         Vector2 setVec = new Vector2(player.AttackRange * dir, 0);
